refactor(expr): move expressoin test to function-v2
