+++ src/query/codegen/src/writes/arithmetics_type_v2.rs
@@ -30,7 +30,7 @@ pub enum OP {
 }
 
 pub fn codegen_arithmetic_type_v2() {
-    let dest = Path::new("src/query/expression/src/types");
+    let dest = Path::new("src/query/expression/src/utils");
     let path = dest.join("arithmetics_type.rs");
 
     let mut file = File::create(&path).expect("open");
