+++ src/query/expression/src/lib.rs
@@ -31,12 +31,7 @@
 #[allow(dead_code)]
 mod chunk;
 
-mod column_from;
-
 pub mod converts;
-pub mod date_helper;
-mod display;
-mod error;
 mod evaluator;
 mod expression;
 mod function;
@@ -45,16 +40,15 @@ mod property;
 mod register;
 pub mod type_check;
 pub mod types;
-pub mod util;
+pub mod utils;
 pub mod values;
 
 pub use crate::chunk::*;
-pub use crate::column_from::*;
-pub use crate::error::*;
 pub use crate::evaluator::*;
 pub use crate::expression::*;
 pub use crate::function::*;
 pub use crate::kernels::*;
 pub use crate::property::*;
 pub use crate::register::*;
+pub use crate::utils::*;
 pub use crate::values::*;
