+++ src/query/expression/src/utils/arrow.rs
@@ -14,7 +14,6 @@
 
 use std::io::Cursor;
 
-use chrono_tz::Tz;
 use common_arrow::arrow::array::Array;
 use common_arrow::arrow::bitmap::Bitmap;
 use common_arrow::arrow::bitmap::MutableBitmap;
@@ -25,27 +24,6 @@ use common_arrow::arrow::io::ipc::read::read_file_metadata;
 use common_arrow::arrow::io::ipc::read::FileReader;
 use common_arrow::arrow::io::ipc::write::FileWriter;
 use common_arrow::arrow::io::ipc::write::WriteOptions;
-
-use crate::error::Result;
-use crate::types::AnyType;
-use crate::types::DataType;
-use crate::Chunk;
-use crate::Column;
-use crate::Evaluator;
-use crate::FunctionRegistry;
-use crate::RawExpr;
-use crate::Span;
-use crate::Value;
-
-pub fn column_merge_validity(column: &Column, bitmap: Option<Bitmap>) -> Option<Bitmap> {
-    match column {
-        Column::Nullable(c) => match bitmap {
-            None => Some(c.validity.clone()),
-            Some(v) => Some(&c.validity & (&v)),
-        },
-        _ => bitmap,
-    }
-}
 
 pub fn bitmap_into_mut(bitmap: Bitmap) -> MutableBitmap {
     bitmap
@@ -101,48 +79,4 @@ pub fn deserialize_arrow_array(bytes: &[u8]) -> Option<Box<dyn Array>> {
     let mut reader = FileReader::new(cursor, metadata, None, None);
     let col = reader.next()?.ok()?.into_arrays().remove(0);
     Some(col)
-}
-
-pub const fn concat_array<T, const A: usize, const B: usize>(a: &[T; A], b: &[T; B]) -> [T; A + B] {
-    let mut result = std::mem::MaybeUninit::uninit();
-    let dest = result.as_mut_ptr() as *mut T;
-    unsafe {
-        std::ptr::copy_nonoverlapping(a.as_ptr(), dest, A);
-        std::ptr::copy_nonoverlapping(b.as_ptr(), dest.add(A), B);
-        result.assume_init()
-    }
-}
-
-/// A convenient shortcut to evaluate a scalar function.
-pub fn eval_function(
-    span: Span,
-    fn_name: &str,
-    args: impl Iterator<Item = (Value<AnyType>, DataType)>,
-    tz: Tz,
-    num_rows: usize,
-    fn_registry: &FunctionRegistry,
-) -> Result<(Value<AnyType>, DataType)> {
-    let (args, cols) = args
-        .enumerate()
-        .map(|(id, (val, ty))| {
-            (
-                RawExpr::ColumnRef {
-                    span: span.clone(),
-                    id,
-                    data_type: ty.clone(),
-                },
-                (val, ty),
-            )
-        })
-        .unzip();
-    let raw_expr = RawExpr::FunctionCall {
-        span,
-        name: fn_name.to_string(),
-        params: vec![],
-        args,
-    };
-    let (expr, ty) = crate::type_check::check(&raw_expr, fn_registry)?;
-    let chunk = Chunk::new(cols, num_rows);
-    let evaluator = Evaluator::new(&chunk, tz);
-    Ok((evaluator.run(&expr)?, ty))
 }
