+++ src/query/expression/tests/it/main.rs
@@ -16,7 +16,5 @@
 #![feature(try_blocks)]
 
 mod convert;
-// TODO(andylokandy): migrate to crate funciton-v2
-mod expression;
 mod kernel;
 mod parser;
