--- src/query/expression/tests/it/testdata/eval-fail.txt
@@ -1,32 +1,0 @@
-error: 
-  --> SQL:1:1
-  |
-1 | get(create_array(1, 2), 2)
-  | ^^^^^^^^^^^^^^^^^^^^^^^^^^ index out of bounds: the len is 2 but the index is 2
-
-
-
-error: 
-  --> SQL:1:1
-  |
-1 | get(create_array(a, b), idx)
-  | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^ index out of bounds: the len is 2 but the index is 2
-
-
-
-error: 
-  --> SQL:1:1
-  |
-1 | CAST(a AS UINT16)
-  | ^^^^^^^^^^^^^^^^^ unable to cast -4 to UInt16
-
-
-
-error: 
-  --> SQL:1:1
-  |
-1 | CAST(c AS INT16)
-  | ^^^^^^^^^^^^^^^^ unable to cast 11111111111 to Int16
-
-
-
