--- src/query/expression/tests/it/testdata/tyck-fail.txt
@@ -1,74 +1,0 @@
-error: 
-  --> SQL:1:6
-  |
-1 | true AND 1
-  |      ^^^ no overload satisfies `and(Boolean, UInt8)`
-
-has tried possible overloads:
-  and(NULL, Boolean NULL) :: NULL                  : unable to unify `Boolean` with `NULL`
-  and(Boolean NULL, NULL) :: NULL                  : unable to unify `UInt8` with `NULL`
-  and(Boolean, Boolean) :: Boolean                 : unable to unify `UInt8` with `Boolean`
-  and(Boolean NULL, Boolean NULL) :: Boolean NULL  : unable to unify `UInt8` with `Boolean`
-
-
-
-error: 
-  --> SQL:1:5
-  |
-1 | NOT NOT 'a'
-  |     ^^^ no overload satisfies `not(String)`
-
-has tried possible overloads:
-  not(NULL) :: NULL                  : unable to unify `String` with `NULL`
-  not(Boolean) :: Boolean            : unable to unify `String` with `Boolean`
-  not(Boolean NULL) :: Boolean NULL  : unable to unify `String` with `Boolean`
-
-
-
-error: 
-  --> SQL:1:1
-  |
-1 | least(1, 2, 3, a)
-  | ^^^^^^^^^^^^^^^^^ no overload satisfies `least(UInt8, UInt8, UInt8, Boolean)`
-
-has tried possible overloads:
-  least(Int16, Int16, Int16, Int16) :: Int16  : unable to unify `Boolean` with `Int16`
-
-
-
-error: 
-  --> SQL:1:5
-  |
-1 | 'a' and 1
-  |     ^^^ no overload satisfies `and(String, UInt8)`
-
-has tried possible overloads:
-  and(NULL, Boolean NULL) :: NULL                  : unable to unify `String` with `NULL`
-  and(Boolean NULL, NULL) :: NULL                  : unable to unify `String` with `Boolean`
-  and(Boolean, Boolean) :: Boolean                 : unable to unify `String` with `Boolean`
-  and(Boolean NULL, Boolean NULL) :: Boolean NULL  : unable to unify `String` with `Boolean`
-
-
-
-error: 
-  --> SQL:1:1
-  |
-1 | get(create_array(1, 2), 'a')
-  | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^ no overload satisfies `get(Array(UInt8), String)`
-
-has tried possible overloads:
-  get(NULL, Int16 NULL) :: NULL               : unable to unify `Array(UInt8)` with `NULL`
-  get(Array(T0) NULL, NULL) :: NULL           : unable to unify `String` with `NULL`
-  get(Array(T0), Int16) :: T0                 : unable to unify `String` with `Int16`
-  get(Array(T0) NULL, Int16 NULL) :: T0 NULL  : unable to unify `String` with `Int16`
-
-
-
-error: 
-  --> SQL:1:1
-  |
-1 | get_tuple(1)(create_tuple(true))
-  | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ no overload satisfies `get_tuple(1)((Boolean,))`
-
-
-
