+++ src/query/functions-v2/src/aggregates/adaptors/aggregate_null_variadic_adaptor.rs
@@ -20,7 +20,7 @@ use bytes::BytesMut;
 use common_arrow::arrow::bitmap::Bitmap;
 use common_exception::Result;
 use common_expression::types::DataType;
-use common_expression::util::column_merge_validity;
+use common_expression::utils::column_merge_validity;
 use common_expression::Column;
 use common_expression::ColumnBuilder;
 use common_io::prelude::BinaryWriteBuf;
