+++ src/query/functions-v2/src/aggregates/aggregate_sum.rs
@@ -21,13 +21,13 @@ use bytes::BytesMut;
 use common_arrow::arrow::bitmap::Bitmap;
 use common_exception::ErrorCode;
 use common_exception::Result;
-use common_expression::types::arithmetics_type::ResultTypeOfUnary;
 use common_expression::types::number::Number;
 use common_expression::types::ArgType;
 use common_expression::types::DataType;
 use common_expression::types::NumberDataType;
 use common_expression::types::NumberType;
 use common_expression::types::ValueType;
+use common_expression::utils::arithmetics_type::ResultTypeOfUnary;
 use common_expression::with_number_mapped_type;
 use common_expression::Column;
 use common_expression::ColumnBuilder;
